Refactor: Styles get own files, generate columns, init Sidebarbtn
Init SidebarBtn for Home
