+++ src/components/buttons/index.js
@@ -1,0 +1,5 @@
+import SidebarBtn from "./SidebarBtn";
+
+export {
+  SidebarBtn,
+};
